functions.py moved inside src
